Update custom button, remove unneeded file and change file name
