+++ interfaces/Schedule.ts
@@ -1,4 +1,4 @@
-import { definitions } from '../types/database'
+import { definitions } from '@types'
 
 export type ScheduleTableRow = Record<string, definitions['employees'] | number> & {
   employee: definitions['employees']
